Changed the click function of the create button operation.
Previous code required the class .summernote and if there were multiple editors would always add to the first editor. This code goes to the top of the editor in the DOM and targets the preceding element with the data-editor attribute of "summernote" which will be the correct target even if there are multiple editors.